% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnation.R
\name{agnation}
\alias{agnation}
\title{Gets 'agnation' from a systemic functional network input.}
\usage{
agnation(df, sorted = FALSE)
}
\arguments{
\item{df}{The dataframe with a initial column 'features'.}

\item{sorted}{Logical. FALSE by default. If TRUE, agnation arrays increasingly}
}
\value{
The agnation of each feature in the system from dataframe row
}
\description{
This function gets agnation from input systems
}
\examples{
agnation(mood_spanish)
features$features       agnacion
1 menor                    2
2 major                    2
3 indicative               2
4 imperative               2

}
