% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delicacy_sist.R
\name{delicacy_sist}
\alias{delicacy_sist}
\title{Gets maximun delicacy level from a full systemic functional network input}
\usage{
delicacy_sist(max_level)
}
\arguments{
\item{df}{The dataframe with a initial column 'features'.}
}
\value{
The maximun delicacy (ie maximum level) of each feature in the system
}
\description{
This function gets maximum delicacy from input systems
}
\examples{
delicacy_sist(df$features)

}
