% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delicacy_subsist.R
\name{delicacy_subsist}
\alias{delicacy_subsist}
\title{Gets delicacy for each level from a  systemic functional network input}
\usage{
delicacy_subsist(level)
}
\arguments{
\item{df}{The dataframe with a initial column 'features'.}
}
\value{
The  delicacy (ie level) of each feature in the system
}
\description{
This function gets delicacy for each level from input systems
}
\examples{
delicacy_subsist(df$features)

}
