% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agnation_int.R
\name{.agnation_int}
\alias{.agnation_int}
\title{Gets 'agnation' from a systemic functional network input}
\usage{
.agnation_int(opositions)
}
\arguments{
\item{df}{The dataframe with a initial column 'features'.}
}
\value{
The agnation of each feature in the system from dataframe row
}
\description{
This function gets agnation from input systems
}
\examples{
agnation(df$features)

}
