% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatures.R
\name{getFeatures}
\alias{getFeatures}
\title{Gets all singular features from a systemic functional network input}
\usage{
getFeatures(df, single = TRUE)
}
\arguments{
\item{df}{The dataframe with a initial column named 'features'.}
}
\value{
The features of each complex system structure in the 'features' column
}
\description{
This function gets all the features from input systems. Since input systems are complex in their structure, this functions clears opositions and lower delicacy levels. It requires `stringr` library
}
\examples{

}
