% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeur_subsist.R
\name{valeur_subsist}
\alias{valeur_subsist}
\title{Gets valeur for each level (ie subsystem) from a  systemic functional network input}
\usage{
valeur_subsist(a, d)
}
\arguments{
\item{df}{The dataframe with a initial column 'features'.}
}
\value{
The  valeur (ie delicacy x agnation) of each feature in the system
}
\description{
This function gets valeur for each level (ie subsystem) from input systems
}
\examples{
valeur_subsist(df$features)

}
