% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeurCalc.R
\name{valeurCalc}
\alias{valeurCalc}
\title{Gets 'valeur' from a systemic functional network input}
\usage{
valeurCalc(df)
}
\arguments{
\item{df}{The dataframe with a initial column 'features'.}
}
\value{
The valeur of each feature in the system
}
\description{
This full function gets vavaleurlue from input systems. Also, gets delicacy subsystem, max delicacy (level), agnation, and valeur subsystem.
}
\examples{

}
