% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valeurCalc2.R
\name{valeurCalc2}
\alias{valeurCalc2}
\title{Gets 'valeur' from a systemic functional network input}
\usage{
valeurCalc2(df, sorted = FALSE)
}
\arguments{
\item{df}{The dataframe with a initial column named 'features'.}

\item{sorted}{Logical. FALSE by default. If TRUE, valeur_feature arrays increasingly}
}
\value{
The valeur of each feature in the system
}
\description{
This 2nd version full function gets valeur from input systems. Also, gets delicacy, agnation, feature's valeur and subsystem's valeur. It requires `stringr` package
}
\examples{

}
